fix 1 so bug up date ngay 26/7

+++ Project/app/Http/Controllers/ProductController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Models\Category;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -23,8 +24,11 @@ class ProductController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Product $product)
     {
+    
+        $this->authorize('viewAny', Product::class);
+       
         $products = Product::latest()->paginate(5);
         return view('admin.products.index', compact('products'));
     }
@@ -36,10 +40,13 @@ class ProductController extends Controller
      */
     public function create(Product $product)
     {
-
-        $categories= Category::all();
-        return view('admin.products.add',compact('categories'));
-
+    //    dd(Auth::user());
+       if( $this->authorize('create', Product::class)){
+        $categories = Category::all();
+        return view('admin.products.add', compact('categories'));
+       }else{
+        abort(403);
+       }
     }
 
     /**
@@ -48,36 +55,37 @@ class ProductController extends Controller
      * @param  \App\Http\Requests\StoreProductRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreProductRequest $request,Product $product)
+    public function store(StoreProductRequest $request, Product $product)
     {
-
-        $product = new Product();
-        $product->name = $request->name;
-        $product->price = $request->price;
-        $product->amouth = $request->amouth;
-        $product->use = $request->use;
-        $product->status = $request->status;
-        $product->description = $request->description;
-        $product->category_id = $request->category_id;
-        
-        if ($request->hasFile('file')) {
-            $file = $request->file;
-            $fileExtension = $file->getClientOriginalExtension();//jpg,png lấy ra định dạng file và trả về
-            $fileName = time();//45678908766 tạo tên file theo thời gian
-            $newFileName = $fileName.'.'.$fileExtension;//45678908766.jpg
-            // $product->image = $newFileName;// cột image gán bằng tên file mới
-            $request->file('file')->storeAs('public/images', $newFileName);//lưu file vào mục public/images với tê mới là $newFileName
-            $product->image = $newFileName;
-
-        } 
-
+        $this->authorize('create', Product::class);
         try {
+            $product = new Product();
+            $product->name = $request->name;
+            $product->price = $request->price;
+            $product->amouth = $request->amouth;
+            $product->use = $request->use;
+            $product->status = $request->status;
+            $product->description = $request->description;
+            $product->category_id = $request->category_id;
+            $product->user_id = auth()->user()->id;
+
+            if ($request->hasFile('file')) {
+                $file = $request->file;
+                $fileExtension = $file->getClientOriginalExtension(); //jpg,png lấy ra định dạng file và trả về
+                $fileName = time(); //45678908766 tạo tên file theo thời gian
+                $newFileName = $fileName . '.' . $fileExtension; //45678908766.jpg
+                // $product->image = $newFileName;// cột image gán bằng tên file mới
+                $request->file('file')->storeAs('public/images', $newFileName); //lưu file vào mục public/images với tê mới là $newFileName
+                $product->image = $newFileName;
+            }
+
+
             $product->save();
             Session::flash('success', 'Thêm Sản phẩm thành công');
             return redirect()->route('product.index');
         } catch (Exception $e) {
             Session::flash('success', 'Thêm Sản phẩm  khong thành công');
-            Log::eror('error'.$e->getMessage().'line________'.$e->getLine());
+            Log::error('error' . $e->getMessage() . 'line________' . $e->getLine());
             return redirect()->back();
         }
     }
@@ -88,10 +96,12 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
+
+
     public function show(Product $product)
     {
-        // $product =  Product::find($id);
-       
+        $this->authorize('view', Product::class);
+
         return view('admin.products.show', compact('product'));
     }
 
@@ -101,12 +111,15 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
+
+
     public function edit(Product $product)
     {
-        $id= $product->id;
-       
-        $categories= Category::all();
-        return view('admin.products.edit',compact('id','product','categories'));
+        $this->authorize('update', Product::class);
+        $id = $product->id;
+
+        $categories = Category::all();
+        return view('admin.products.edit', compact('id', 'product', 'categories'));
     }
 
     /**
@@ -116,34 +129,37 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(StoreProductRequest $request, Product $product)
+
+
+    public function update(UpdateProductRequest $request, Product $product)
     {
-
-        $product->name = $request->name;
-        $product->price = $request->price;
-        $product->amouth = $request->amouth;
-        $product->use = $request->use;
-        $product->status = $request->status;
-        $product->description = $request->description;
-        $product->category_id = $request->category_id;
-        
-        if ($request->hasFile('file')) {
-            $file = $request->file;
-            $fileExtension = $file->getClientOriginalExtension();//jpg,png lấy ra định dạng file và trả về
-            $fileName = time();//45678908766 tạo tên file theo thời gian
-            $newFileName = $fileName.'.'.$fileExtension;//45678908766.jpg
-            // $product->image = $newFileName;// cột image gán bằng tên file mới
-            $request->file('file')->storeAs('public/images', $newFileName);//lưu file vào mục public/images với tê mới là $newFileName
-            $product->image = $newFileName;
-
-        }
-
+        $this->authorize('update', Product::class);
         try {
+            $product->name = $request->name;
+            $product->price = $request->price;
+            $product->amouth = $request->amouth;
+            $product->use = $request->use;
+            $product->status = $request->status;
+            $product->description = $request->description;
+            $product->category_id = $request->category_id;
+
+            if ($request->hasFile('file')) {
+                $file = $request->file;
+                $fileExtension = $file->getClientOriginalExtension(); //jpg,png lấy ra định dạng file và trả về
+                $fileName = time(); //45678908766 tạo tên file theo thời gian
+                $newFileName = $fileName . '.' . $fileExtension; //45678908766.jpg
+                $request->file('file')->storeAs('public/images', $newFileName); //lưu file vào mục public/images với tê mới là $newFileName
+                $product->image = $newFileName; // cột image gán bằng tên file mới
+            }else{
+                $product->image = $product->image;
+            }
             $product->save();
             Session::flash('success', 'chỉnh sửa thành công');
             return redirect()->route('product.index');
         } catch (Exception $e) {
-            echo 'lỗi' . $e->getMessage();
+            Session::flash('error', 'chỉnh sửa  thất bại');
+            Log::eror('error update' . $e->getMessage() . 'line________' . $e->getLine());
+            return redirect()->back();
         }
     }
 
@@ -153,26 +169,27 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
-    
+
+
     public function destroy($id)
-    { 
+    {
+        $this->authorize('delete', Product::class);
         try {
+
             DB::beginTransaction();
             $product = Product::findOrFail($id);
-        $product->delete();
-        DB::commit();
-        Session::flash('success', 'Xóa sản phẩm thành công');
-        return redirect()->back();  
+            $product->delete();
+            DB::commit();
+            Session::flash('success', 'Xóa sản phẩm thành công');
+            return redirect()->back();
         } catch (\Exception $e) {
             DB::rollBack();
             Session::flash('errors', 'Xóa danh mục vĩnh viễn lỗi!!! Hãy thử lại');
-            Log::error('messages' . $e->getMessage() . 'line________'.$e->getLine());
+            Log::error('messages' . $e->getMessage() . 'line________' . $e->getLine());
             return redirect()->route('category.index');
-            
         }
-      
-
     }
+
     public function trashed()
     {
         $producs_trasheds = Product::onlyTrashed()->get();
@@ -181,11 +198,13 @@ class ProductController extends Controller
 
     public function restore($id)
     {
+        $this->authorize('delete', Product::class);
+
         try {
             DB::beginTransaction();
             $category = Product::withTrashed()->where('id', $id)->restore();
             DB::commit();
-        Session::flash('success', 'Phục hồi danh mục thành công');
+            Session::flash('success', 'Phục hồi danh mục thành công');
             return redirect()->route('product.index');
         } catch (\Exception $e) {
             DB::rollBack();
@@ -197,9 +216,11 @@ class ProductController extends Controller
 
     public function forceDelete($id)
     {
+        $this->authorize('forceDelete', Product::class);
+
         try {
             DB::beginTransaction();
-            $product= Product::withTrashed()->find($id);
+            $product = Product::withTrashed()->find($id);
             // $product_categories=$category->products;
             // foreach ($product_categories as $key => $product_category) {
             // $product_category->category_id= null;
@@ -207,19 +228,14 @@ class ProductController extends Controller
             // dd($product_category);
             // }
             $product->forceDelete();
-        DB::commit();
-        Session::flash('success', 'Xóa danh mục vĩnh viễn thành công');
-        return redirect()->route('product-trashed');
+            DB::commit();
+            Session::flash('success', 'Xóa danh mục vĩnh viễn thành công');
+            return redirect()->route('product-trashed');
         } catch (\Exception $e) {
             DB::rollBack();
-        Session::flash('errors', 'Xóa danh mục vĩnh viễn lỗi!!! Hãy thử lại');
-        Log::error('messages' . $e->getMessage() . 'line________'.$e->getLine());
-        return redirect()->route('product-trashed');
-
-
-            
+            Session::flash('errors', 'Xóa danh mục vĩnh viễn lỗi!!! Hãy thử lại');
+            Log::error('messages' . $e->getMessage() . 'line________' . $e->getLine());
+            return redirect()->route('product-trashed');
         }
-        
-        
     }
 }
